+++ example/router.rb
@@ -7,6 +7,10 @@ module EventHub
       "1.0.0"
     end
 
+    def http_resources
+      [:heartbeat, :version, :docs, :changelog, :configuration]
+    end
+
     def handle_message(message, args = {})
       id = message.body["id"]
       EventHub.logger.info("Received: [#{id}]")